requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow
colorthief
tqdm
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
sqlite-vec
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
numpy
aiofiles==23.2.1
sqlalchemy
jwt
sqlalchemy
weaviate-client==4.6.2
psycopg2-binary==2.9.9
python-jose

